Rename rule to cyclomatic-complexity
