+++ eslint-bridge/src/rules/cyclomatic-complexity.ts
