+++ eslint-bridge/src/rules/main.ts
@@ -33,7 +33,7 @@ import { rule as maxUnionSize } from "./max-union-size";
 import { rule as noDuplicateInComposite } from "./no-duplicate-in-composite";
 import { rule as noInconsistentReturns } from "./no-inconsistent-returns";
 import { rule as noRedundantJump } from "./no-redundant-jump";
-import { rule as mcCabeRule } from "./mccabe-complexity";
+import { rule as cyclomaticComplexity } from "./cyclomatic-complexity";
 import { rule as noTab } from "./no-tab";
 import { rule as nonExistentOperator } from "./non-existent-operator";
 import { rule as noNestedSwitch } from "./no-nested-switch";
@@ -73,7 +73,7 @@ ruleModules["max-union-size"] = maxUnionSize;
 ruleModules["no-duplicate-in-composite"] = noDuplicateInComposite;
 ruleModules["no-inconsistent-returns"] = noInconsistentReturns;
 ruleModules["no-redundant-jump"] = noRedundantJump;
-ruleModules["mccabe-complexity"] = mcCabeRule;
+ruleModules["cyclomatic-complexity"] = cyclomaticComplexity;
 ruleModules["no-tab"] = noTab;
 ruleModules["non-existent-operator"] = nonExistentOperator;
 ruleModules["no-nested-switch"] = noNestedSwitch;
