+++ javascript-checks/src/main/java/org/sonar/javascript/checks/CheckList.java
@@ -96,6 +96,7 @@ public final class CheckList {
       CookiesCheck.class,
       CorsCheck.class,
       CounterUpdatedInLoopCheck.class,
+      CyclomaticComplexityCheck.class,
       DeadStoreCheck.class,
       DebuggerStatementCheck.class,
       DeclarationInGlobalScopeCheck.class,
@@ -162,7 +163,6 @@ public final class CheckList {
       LineLengthCheck.class,
       LocalStorageCheck.class,
       LoopsShouldNotBeInfiniteCheck.class,
-      McCabeComplexityCheck.class,
       MaxSwitchCasesCheck.class,
       MaxUnionSizeCheck.class,
       MisorderedParameterListCheck.class,
